LayoutTests/fast/canvas/webgl/webgl-array-invalid-ranges-expected.txt

Verifies that out-of-range parameters for creation, slicing and setting of WebGL arrays are caught

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS Construction of WebGLByteArray with null buffer threw exception
PASS Construction of WebGLByteArray with out-of-range values threw exception
PASS Construction of WebGLByteArray with negative out-of-range values threw exception
Testing slicing of WebGLByteArray
PASS array.length is 32 / typeSize
PASS array.slice(4, 0x3FFFFFFF).length is (32 / typeSize) - 4
PASS array.slice(4, -2147483648).length is 0
PASS Setting WebGLByteArray from array with out-of-range offset was caught
PASS Setting WebGLByteArray from fake array with invalid length was caught
PASS Setting WebGLByteArray from WebGLByteArray with out-of-range offset was caught
PASS Construction of WebGLUnsignedByteArray with null buffer threw exception
PASS Construction of WebGLUnsignedByteArray with out-of-range values threw exception
PASS Construction of WebGLUnsignedByteArray with negative out-of-range values threw exception
Testing slicing of WebGLUnsignedByteArray
PASS array.length is 32 / typeSize
PASS array.slice(4, 0x3FFFFFFF).length is (32 / typeSize) - 4
PASS array.slice(4, -2147483648).length is 0
PASS Setting WebGLUnsignedByteArray from array with out-of-range offset was caught
PASS Setting WebGLUnsignedByteArray from fake array with invalid length was caught
PASS Setting WebGLUnsignedByteArray from WebGLUnsignedByteArray with out-of-range offset was caught
PASS Construction of WebGLShortArray with null buffer threw exception
PASS Construction of WebGLShortArray with out-of-range values threw exception
PASS Construction of WebGLShortArray with negative out-of-range values threw exception
PASS Construction of huge WebGLShortArray threw exception
Testing slicing of WebGLShortArray
PASS array.length is 32 / typeSize
PASS array.slice(4, 0x3FFFFFFF).length is (32 / typeSize) - 4
PASS array.slice(4, -2147483648).length is 0
PASS Setting WebGLShortArray from array with out-of-range offset was caught
PASS Setting WebGLShortArray from fake array with invalid length was caught
PASS Setting WebGLShortArray from WebGLShortArray with out-of-range offset was caught
PASS Construction of WebGLUnsignedShortArray with null buffer threw exception
PASS Construction of WebGLUnsignedShortArray with out-of-range values threw exception
PASS Construction of WebGLUnsignedShortArray with negative out-of-range values threw exception
PASS Construction of huge WebGLUnsignedShortArray threw exception
Testing slicing of WebGLUnsignedShortArray
PASS array.length is 32 / typeSize
PASS array.slice(4, 0x3FFFFFFF).length is (32 / typeSize) - 4
PASS array.slice(4, -2147483648).length is 0
PASS Setting WebGLUnsignedShortArray from array with out-of-range offset was caught
PASS Setting WebGLUnsignedShortArray from fake array with invalid length was caught
PASS Setting WebGLUnsignedShortArray from WebGLUnsignedShortArray with out-of-range offset was caught
PASS Construction of WebGLIntArray with null buffer threw exception
PASS Construction of WebGLIntArray with out-of-range values threw exception
PASS Construction of WebGLIntArray with negative out-of-range values threw exception
PASS Construction of huge WebGLIntArray threw exception
Testing slicing of WebGLIntArray
PASS array.length is 32 / typeSize
PASS array.slice(4, 0x3FFFFFFF).length is (32 / typeSize) - 4
PASS array.slice(4, -2147483648).length is 0
PASS Setting WebGLIntArray from array with out-of-range offset was caught
PASS Setting WebGLIntArray from fake array with invalid length was caught
PASS Setting WebGLIntArray from WebGLIntArray with out-of-range offset was caught
PASS Construction of WebGLUnsignedIntArray with null buffer threw exception
PASS Construction of WebGLUnsignedIntArray with out-of-range values threw exception
PASS Construction of WebGLUnsignedIntArray with negative out-of-range values threw exception
PASS Construction of huge WebGLUnsignedIntArray threw exception
Testing slicing of WebGLUnsignedIntArray
PASS array.length is 32 / typeSize
PASS array.slice(4, 0x3FFFFFFF).length is (32 / typeSize) - 4
PASS array.slice(4, -2147483648).length is 0
PASS Setting WebGLUnsignedIntArray from array with out-of-range offset was caught
PASS Setting WebGLUnsignedIntArray from fake array with invalid length was caught
PASS Setting WebGLUnsignedIntArray from WebGLUnsignedIntArray with out-of-range offset was caught
PASS Construction of WebGLFloatArray with null buffer threw exception
PASS Construction of WebGLFloatArray with out-of-range values threw exception
PASS Construction of WebGLFloatArray with negative out-of-range values threw exception
PASS Construction of huge WebGLFloatArray threw exception
Testing slicing of WebGLFloatArray
PASS array.length is 32 / typeSize
PASS array.slice(4, 0x3FFFFFFF).length is (32 / typeSize) - 4
PASS array.slice(4, -2147483648).length is 0
PASS Setting WebGLFloatArray from array with out-of-range offset was caught
PASS Setting WebGLFloatArray from fake array with invalid length was caught
PASS Setting WebGLFloatArray from WebGLFloatArray with out-of-range offset was caught
PASS successfullyParsed is true

TEST COMPLETE
